requirements.txt
pandas==3.0.1
langchain-core==1.2.21
langchain-community==0.4.1
langchain-text-splitters==1.1.1
langchain-huggingface==1.2.1
langchain-groq==1.1.2
python-dotenv==1.0.1
sentence-transformers==5.3.0
chromadb==0.5.20